backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# AI and ML
openai-whisper==20231117
groq==0.4.2

# Text-to-Speech
edge-tts==6.1.10

# Audio/Video processing
ffmpeg-python==0.2.0
pydub==0.25.1

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
aiofiles==23.2.1

# CORS
httpx==0.26.0
